+++ source/include/Ostap/Peaks.h
@@ -3349,11 +3349,18 @@ namespace Ostap
       /// get the (excess) kurtosis 
       double kurtosis () const ;
       // ======================================================================
+    public: // support 
+      // ======================================================================
+      /// x-min
+      double xmin () const { return m_mu - m_varsigma ; }
+      /// x-max
+      double xmax () const { return m_mu + m_varsigma ; }
+      // ======================================================================
     public: // integrals 
       // ======================================================================
       /// integral 
       double integral () const ;
-      /// integral from 
+      /// integral from low to high 
       double integral
       ( const double low  , 
         const double high ) const ;
@@ -3445,6 +3452,13 @@ namespace Ostap
       /// get the (excess) kurtosis 
       double kurtosis () const ;
       // ======================================================================
+    public: // support 
+      // ======================================================================
+      /// x-min
+      double xmin () const { return m_mu - m_varsigma ; }
+      /// x-max
+      double xmax () const { return m_mu + m_varsigma ; }
+      // ======================================================================
     public: // integrals 
       // ======================================================================
       /// integral 
@@ -3475,6 +3489,103 @@ namespace Ostap
       double m_mu        { 0 } ;
       /// scale paramewter 
       double m_varsigma  { 0 } ;
+      // ====================================================================== 
+    private:
+      // ======================================================================
+      /// integration workspace
+      Ostap::Math::WorkSpace m_workspace {} ; // integration workspace
+      // ======================================================================      
+    } ;
+    // ========================================================================
+    /** @class FupM 
+     *  Finite stomic functon <code>fup_N</code>,  a finite solution
+     *  of the equation 
+     *  \f[ f^{\prime(x) = 2 \left( f( 2x+1) - f(2x-1)\right) }\f] with
+     *  \f$ f(0) = 1 \f$ 
+     *  @see Ostap::Math::fupN_F 
+     */
+    class FupN
+    {
+    public:
+      // ======================================================================
+      /// constructor with location and scale parmaeters 
+      FupN 
+      ( const unsigned short N  = 1 , 
+        const double   mu       = 0 , 
+        const double   varsigma = 1 ) ;
+      // ======================================================================
+    public:
+      // ======================================================================
+      /// evaluate the function 
+      double evaluate ( const double x ) const ;
+      /// evaluate the function 
+      inline double pdf         ( const double x ) const 
+      { return evaluate ( x )  ; }
+      /// evaluate the function 
+      inline double operator()  ( const double x ) const 
+      { return evaluate ( x )  ; }
+      // ======================================================================
+    public: // getters 
+      // ======================================================================
+      /// get N 
+      unsigned short N        () const { return m_N        ; }
+      /// get mu 
+      double         mu       () const { return m_mu       ; }
+      /// get varsigma 
+      double         varsigma () const { return m_varsigma ; } 
+      // ======================================================================
+    public: // setters 
+      // ======================================================================
+      /// set mu
+      bool setMu        ( const double value ) ;
+      /// set varsigma 
+      bool setVarsigma  ( const double value ) ;
+      // ======================================================================
+    public: // some properteis 
+      // ======================================================================
+      /// get the mean     of the distribution 
+      double mean     () const { return m_mu ; }      
+      /// get the mode     of the distribution 
+      double mode     () const { return m_mu ; }      
+      /// get the median   of the distribution 
+      double median   () const { return m_mu ; }      
+      /// get the skewness 
+      double skewness () const { return 0    ; }
+      // ======================================================================
+    public: // support 
+      // ======================================================================
+      /// x-min
+      double xmin () const { return m_mu - 0.5 * ( m_N + 2 ) * m_varsigma ; }
+      /// x-max
+      double xmax () const { return m_mu + 0.5 * ( m_N + 2 ) * m_varsigma ; }
+      // ======================================================================
+    public: // integrals 
+      // ======================================================================
+      /// integral 
+      double integral () const ;
+      /// integral from low to high  
+      double integral
+      ( const double low  , 
+        const double high ) const ;
+      // ======================================================================
+    public: // unique tag 
+      // ======================================================================
+      /// unique tag 
+      std::size_t tag() const ;
+      // ======================================================================
+    private:
+      // ======================================================================
+      /// evaluate the "standard" fupN function 
+      double eval ( const double z )  const ;
+      // ======================================================================
+    private:
+      // ======================================================================
+      /// parameter N 
+      unsigned short m_N         { 1 } ;
+      /// location parametyer 
+      double         m_mu        { 0 } ;
+      /// scale paramewter 
+      double         m_varsigma  { 1 } ;
       // ====================================================================== 
     private:
       // ======================================================================