+++ source/src/PDFs.cpp
@@ -6785,6 +6785,89 @@ Double_t Ostap::Models::Up::analyticalIntegral
 // ============================================================================
 
 
+// ============================================================================
+// constructor from all parameters 
+// ============================================================================
+Ostap::Models::FupN::FupN
+( const char*          name     , 
+  const char*          title    ,
+  RooAbsReal&          x        ,
+  const unsigned short N        ,
+  RooAbsReal&          mu       ,
+  RooAbsReal&          varsigma ) 
+  : RooAbsPdf ( name , title ) 
+    //
+  , m_x        ( "!x"        , "observable" , this , x        ) 
+  , m_mu       ( "!mu"       , "location"   , this , mu       ) 
+  , m_varsigma ( "!varsigma" , "scale"      , this , varsigma ) 
+    //
+  , m_fupN ( N , 0 , 1  ) 
+{
+  setPars() ;
+}
+// ============================================================================
+// "copy" constructor 
+// ============================================================================
+Ostap::Models::FupN::FupN
+( const Ostap::Models::FupN& right ,
+  const char*                name  ) 
+  : RooAbsPdf ( right , name ) 
+    //
+  , m_x         ( "!x"        , this , right.m_x        ) 
+  , m_mu        ( "!mu"       , this , right.m_mu       )
+  , m_varsigma  ( "!varsigma" , this , right.m_varsigma )
+    //
+  , m_fupN ( right.m_fupN ) 
+{
+  setPars () ;
+}
+// ============================================================================
+// destructor
+// ============================================================================
+Ostap::Models::FupN::~FupN(){}
+// ============================================================================
+// clone 
+// ============================================================================
+Ostap::Models::FupN*
+Ostap::Models::FupN::clone( const char* name ) const 
+{ return new Ostap::Models::FupN( *this , name) ; }
+// ============================================================================
+void Ostap::Models::FupN::setPars () const 
+{
+  m_fupN.setMu       ( m_mu       ) ;
+  m_fupN.setVarsigma ( m_varsigma ) ;
+}
+// ============================================================================
+// the actual evaluation of function 
+// ============================================================================
+Double_t Ostap::Models::FupN::evaluate() const 
+{
+  setPars () ;
+  return m_fupN ( m_x ) ;
+}
+// ============================================================================
+Int_t Ostap::Models::FupN::getAnalyticalIntegral
+( RooArgSet&     allVars      , 
+  RooArgSet&     analVars     ,
+  const char* /* rangename */ ) const 
+{
+  if ( matchArgs ( allVars , analVars , m_x ) ) { return 1 ; }
+  return 0 ;
+}
+// ============================================================================
+Double_t Ostap::Models::FupN::analyticalIntegral 
+( Int_t       code      , 
+  const char* rangeName ) const 
+{
+  assert ( code == 1 ) ;
+  if ( 1 != code ) {}
+  //
+  setPars () ;
+  return m_fupN.integral ( m_x.min(rangeName) , m_x.max(rangeName) ) ;
+}
+// ============================================================================
+
+
 
 // ============================================================================
 // constructor from all parameters 
@@ -8383,6 +8466,7 @@ ClassImp(Ostap::Models::Slash              )
 ClassImp(Ostap::Models::AsymmetricLaplace  ) 
 ClassImp(Ostap::Models::Hat                ) 
 ClassImp(Ostap::Models::Up                 ) 
+ClassImp(Ostap::Models::FupN               ) 
 ClassImp(Ostap::Models::Tsallis            ) 
 ClassImp(Ostap::Models::QGSM               ) 
 ClassImp(Ostap::Models::TwoExpos           )